🔄 synced local 'skyvern-frontend/src/' with remote 'skyvern-frontend/src/'
<!-- ELLIPSIS_HIDDEN -->

> [!IMPORTANT]
> Refactor tooltip logic for workflow blocks by centralizing it in a new hook and updating nodes to use this hook for displaying parameter addition tips.
>
>   - **Behavior**:
>     - Removed tooltip logic from `WorkflowBlockInput.tsx` and `WorkflowBlockInputTextarea.tsx`.
>     - Added `useIsFirstBlockInWorkflow` hook in `useIsFirstNodeInWorkflow.tsx` to determine if a node is the first in the workflow.
>     - Updated nodes (`ActionNode.tsx`, `ExtractionNode.tsx`, `FileDownloadNode.tsx`, `FileParserNode.tsx`, `LoginNode.tsx`, `LoopNode.tsx`, `NavigationNode.tsx`, `PDFParserNode.tsx`, `SendEmailNode.tsx`, `TaskNode.tsx`, `TextPromptNode.tsx`, `ValidationNode.tsx`, `WaitNode.tsx`) to use `useIsFirstBlockInWorkflow` for displaying parameter addition tips.
>   - **Code Cleanup**:
>     - Removed `isFirstInputInNode` prop from `WorkflowBlockInput` and `WorkflowBlockInputTextarea` components.
>     - Removed unused imports related to tooltips in `WorkflowBlockInput.tsx` and `WorkflowBlockInputTextarea.tsx`.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for f265b8fb17409c366c5177ecb333a2de341c2018. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
